fix: round 1 of fixes for debugger
- renamed some pretty names of parser classes
- renamed `XAbstractWeakMap` to `XWeakMapImpl` as that made more sense
- added since notes to public APIs
- used `.clear()` for mutable collections like normal people do
- chain `.getOrElse()` in `Rename` as it was less clunky than one match and one `.getOrElse()`

TODO: fix public ParseAttempt with trait instead of case class
