+++ parsley-debug/js/src/main/scala-2.13/parsley/debugger/internal/XWeakMap.scala
@@ -8,7 +8,7 @@ package parsley.debugger.internal
 import scala.collection.mutable
 
 private [parsley] final class XWeakMap[K <: Object, V](startSize: Int = 32) extends mutable.Map[K, V] { // scalastyle:ignore magic.number
-    private [internal] val backing: XAbstractWeakMap[K, V] = new XAbstractWeakMap(startSize)
+    private [internal] val backing: XWeakMapImpl[K, V] = new XWeakMapImpl(startSize)
 
     override def subtractOne(key: K): XWeakMap.this.type = {
         backing.drop(key)
