+++ parsley-debug/js/src/main/scala/parsley/debugger/internal/XWeakMapImpl.scala
@@ -9,13 +9,13 @@ import scala.annotation.tailrec
 import scala.collection.mutable.ListBuffer
 import scala.scalajs.js.WeakRef
 
-import XAbstractWeakMap.Backing
+import XWeakMapImpl.Backing
 import org.typelevel.scalaccompat.annotation.unused
 
 // XXX: This map has a heavy dependency on JS's WeakRef. Not all types will be accepted as keys as
 //      they must qualify as Objects to be inserted as keys into a weak reference.
 //      Simple types like strings will raise a runtime exception.
-private [internal] final class XAbstractWeakMap[K <: Object, V](startSize: Int = 32) { // scalastyle:ignore magic.number
+private [internal] final class XWeakMapImpl[K <: Object, V](startSize: Int = 32) { // scalastyle:ignore magic.number
     // Constants and helpers.
     private val minBuckets: Int = 8
     private val maxBucketConstant: Double = 4.0
@@ -167,6 +167,6 @@ private [internal] final class XAbstractWeakMap[K <: Object, V](startSize: Int =
     }
 }
 
-private object XAbstractWeakMap {
+private object XWeakMapImpl {
     type Backing[K, V] = Array[ListBuffer[(WeakRef[K], V)]]
 }
