+++ parsley-debug/jvm-native/src/main/scala-2.13/parsley/debugger/internal/XWeakMap.scala
@@ -11,7 +11,7 @@ import org.typelevel.scalaccompat.annotation.unused
 
 // For the JVM and Native, its WeakHashMap does the job.
 private [parsley] final class XWeakMap[K, V](@unused startSize: Int = 32) extends mutable.Map[K, V] { // scalastyle:ignore magic.number
-    private val backing: XAbstractWeakMap[K, V] = new XAbstractWeakMap()
+    private val backing: XWeakMapImpl[K, V] = new XWeakMapImpl()
 
     // $COVERAGE-OFF$
     override def subtractOne(k: K): XWeakMap.this.type = {
